Refactors taskpublished to use taskdata
