+++ taskpublisher/src/main/kotlin/io/taskdata/db/mapping.kt
@@ -1,7 +1,7 @@
-package io.taskpublisher.db
+package io.taskdata.db
 
-import io.taskpublisher.Status
-import io.taskpublisher.Task
+import io.taskdata.Status
+import io.taskdata.Task
 import kotlinx.coroutines.Dispatchers
 import org.jetbrains.exposed.dao.Entity
 import org.jetbrains.exposed.dao.EntityClass
